Finishing touch on login

+++ src/app/components/artist/artist.component.ts
@@ -194,16 +194,19 @@ export class ArtistComponent extends Loading implements OnInit {
 
   // Adds the artist as a favourite for the user in the DB
   addFavourite() {
-    this.reviewService.addArtistAsFavourite(this.userInfo.uid, this.artist);
-    this.isFavourite = true;
+    if(this.isAuth()) {
+      this.reviewService.addArtistAsFavourite(this.userInfo.uid, this.artist);
+      this.isFavourite = true;
+    }
   }
 
   removeFavourite() {
-  	this.reviewService.removeArtistAsFavourite(this.userInfo.uid, this.artist);
-    this.isFavourite = false;
+    if(this.isAuth()) {
+    	this.reviewService.removeArtistAsFavourite(this.userInfo.uid, this.artist);
+      this.isFavourite = false;
+    }
   }
 
-  // Global authentication methods
   isAuth() {
     return this.authService.isAuthenticated();
   }
